api/requirements.txt
fastapi[standard]
python-multipart
pillow
numpy
matplotlib
tensorflow==2.12.0
opencv-python
inference-sdk
